@@ -1,7 +1,7 @@
+import 'package:admin/constants.dart';
+import 'package:easy_localization/easy_localization.dart';
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
-
-import '../../../constants.dart';
 
 class Chart extends StatelessWidget {
   const Chart({
@@ -28,14 +28,19 @@ class Chart extends StatelessWidget {
               children: [
                 SizedBox(height: defaultPadding),
                 Text(
-                  "29.1",
+                  '29.1',
                   style: Theme.of(context).textTheme.headline4!.copyWith(
-                        color: Colors.white,
+                        color: Colors.black87,
                         fontWeight: FontWeight.w600,
                         height: 0.5,
                       ),
                 ),
-                Text("of 128GB")
+                Text(
+                  'total'.plural(29.1),
+                  style: TextStyle(
+                    color: Colors.black54,
+                  ),
+                ),
               ],
             ),
           ),
