@@ -1,8 +1,7 @@
+import 'package:admin/constants.dart';
 import 'package:admin/models/MyFiles.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-
-import '../../../constants.dart';
 
 class FileInfoCard extends StatelessWidget {
   const FileInfoCard({
