@@ -1,4 +1,9 @@
+import 'package:admin/constants.dart';
+import 'package:admin/controllers/navigation/navigation_bloc.dart';
+import 'package:admin/models/drawer_listitem.dart';
+import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
 class SideMenu extends StatelessWidget {
@@ -12,48 +17,29 @@ class SideMenu extends StatelessWidget {
       child: ListView(
         children: [
           DrawerHeader(
-            child: Image.asset("assets/images/logo.png"),
+            child: Column(
+              children: [
+                SvgPicture.asset("assets/icons/logo.svg"),
+                SizedBox(height: defaultPadding * 0.5),
+                Text(
+                  tr('charity_name'),
+                  style: TextStyle(
+                    color: Color(0xFF03716E),
+                    fontSize: 14,
+                  ),
+                  textAlign: TextAlign.center,
+                )
+              ],
+            ),
           ),
-          DrawerListTile(
-            title: "Dashboard",
-            svgSrc: "assets/icons/menu_dashbord.svg",
-            press: () {},
-          ),
-          DrawerListTile(
-            title: "Transaction",
-            svgSrc: "assets/icons/menu_tran.svg",
-            press: () {},
-          ),
-          DrawerListTile(
-            title: "Task",
-            svgSrc: "assets/icons/menu_task.svg",
-            press: () {},
-          ),
-          DrawerListTile(
-            title: "Documents",
-            svgSrc: "assets/icons/menu_doc.svg",
-            press: () {},
-          ),
-          DrawerListTile(
-            title: "Store",
-            svgSrc: "assets/icons/menu_store.svg",
-            press: () {},
-          ),
-          DrawerListTile(
-            title: "Notification",
-            svgSrc: "assets/icons/menu_notification.svg",
-            press: () {},
-          ),
-          DrawerListTile(
-            title: "Profile",
-            svgSrc: "assets/icons/menu_profile.svg",
-            press: () {},
-          ),
-          DrawerListTile(
-            title: "Settings",
-            svgSrc: "assets/icons/menu_setting.svg",
-            press: () {},
-          ),
+          for (int index = 0; index < sideMenuOptions.length; index++)
+            DrawerListTile(
+              press: () {
+                BlocProvider.of<NavigationBloc>(context)
+                    .add(NavigateToPage(pageIndex: index));
+              },
+              drawerItem: sideMenuOptions[index],
+            ),
         ],
       ),
     );
@@ -64,12 +50,11 @@ class DrawerListTile extends StatelessWidget {
   const DrawerListTile({
     Key? key,
     // For selecting those three line once press "Command+D"
-    required this.title,
-    required this.svgSrc,
     required this.press,
+    required this.drawerItem,
   }) : super(key: key);
 
-  final String title, svgSrc;
+  final DrawerListItem drawerItem;
   final VoidCallback press;
 
   @override
@@ -78,13 +63,13 @@ class DrawerListTile extends StatelessWidget {
       onTap: press,
       horizontalTitleGap: 0.0,
       leading: SvgPicture.asset(
-        svgSrc,
-        color: Colors.white54,
+        drawerItem.svgSrc,
+        color: Colors.black54,
         height: 16,
       ),
       title: Text(
-        title,
-        style: TextStyle(color: Colors.white54),
+        drawerItem.title,
+        style: TextStyle(color: Colors.black54),
       ),
     );
   }
